xdgportaltest: close tags instead of reopening them
The text sent as notifications are not valid xml; they contain two open
tags instead of an open/close pair. libnotificationmanager handles this
correctly (and there's an autotest to ensure it does), but it will
complain about being handled invalid xml.

This change replaces the second open tag with a close tag.